bpf: encap: allow limited usage from XDP context
The encap helpers are currently enabled if ENCAP_IFINDEX is available. For
XDP we can't redirect to the encap interface, so we don't want visibility
of the actual ifindex. But we still want to call into the common helper
infrastructure from generic code (ie. nodeport.h).

So add a HAVE_ENCAP macro to decide whether we need the encap helpers, and
hide the tunnel-key details in an skb-specific helper. The XDP variant
should not be called by anyone yet.
